refactor: CRUD Livro

+++ src/biblioteca/service/LivroService.java
@@ -1,26 +1,25 @@
 package biblioteca.service;
 
+import java.util.ArrayList;
 import java.util.List;
 
-import biblioteca.entities.Emprestimo;
+import biblioteca.domainException.DadosException;
 import biblioteca.entities.Livro;
-import biblioteca.repository.EmprestimoDAO;
 import biblioteca.repository.LivroDAO;
 
 public class LivroService {
 
 	private LivroDAO livroDao;
-	private EmprestimoDAO emprestimoDao;
 	
 	// injecao de dependencia
-	public LivroService(LivroDAO livroDao, EmprestimoDAO emprestimoDao) {
+	public LivroService(LivroDAO livroDao) {
 		this.livroDao = livroDao;
-		this.emprestimoDao = emprestimoDao;
+		
 	}
 
-
-	public int insert(Livro novoLivro) {
-		List<Livro> livros = livroDao.findAll();
+	public void insert(Livro novoLivro) {
+		List<Livro> livros = new ArrayList<Livro>();
+		livros = livroDao.findAll();
 		for (Livro livro : livros) {
 			if (livro.getTitulo().equals(novoLivro.getTitulo())) {
 				throw new RuntimeException("Livro já existente no catálogo!");
@@ -34,18 +33,36 @@ public class LivroService {
 			novoLivro.setId(ultimoLivro.getId() + 1);
 		}
 		livroDao.insert(novoLivro);
-		return novoLivro.getId();
 	}
 
 	
-	public void delete(int idLivro) {
-		// verifica se tem emprestimo penedente
-		List<Emprestimo> historicoEmprestimo = emprestimoDao.findByIdLivro(idLivro);
-		for (Emprestimo emprestimo : historicoEmprestimo) {
-			if (!emprestimo.isDevolvido()) {
-				throw new RuntimeException("Livro com pendências!");
-			}
+	public Livro findById(int livro_id) {
+		return livroDao.find(livro_id);
+	}
+	
+	public List<Livro> findAll(){
+		List<Livro> livros = livroDao.findAll();
+		if(livros.isEmpty()) {
+			throw new DadosException("Arquivo vazio!");
 		}
+		return livros;
+	}
+	
+	
+	public void update(int id, Livro livro) {
+		livroDao.update(id, livro);
+	}
+	
+	
+	public void delete(int idLivro) {
+		//busca apenas o livro que queremos apagar
+		//Se o ID não existir, o DAO já vai lançar exceção
+		Livro livro = livroDao.find(idLivro);
+		// Verifica se este livro específico está emprestado
+		if(!livro.isDisponivel()) {
+			throw new DadosException("Livro emprestado!");
+		}
+		// Se o livro existe e está na prateleira, pode apagar
 		livroDao.remove(idLivro);
 	}
 }
